+++ packages/extension/example/string.dart
@@ -1,0 +1,7 @@
+import 'package:extension/string.dart';
+
+main() {
+  final capitalized =
+      capitalizeFirstLetter('capitalize my first letter, please!');
+  assert(capitalized == 'Capitalize my first letter, please!');
+}