+++ packages/extension/lib/extension.dart
@@ -1,1 +1,2 @@
 export './enum.dart';
+export './string.dart';