+++ packages/extension/test/all.dart
@@ -1,5 +1,7 @@
 import './enum.dart' as _enum;
+import './string.dart' as _string;
 
 main() {
   _enum.main();
+  _string.main();
 }